chargerpi/requirements.txt
stmpy
paho-mqtt==1.6.0
IPython
tk